backend/requirements.txt
flask==2.0.1
numpy==1.21.0
tensorflow==2.10.0
pillow==9.5.0
scikit-learn==1.0.2
flask-cors==3.0.10
gunicorn==20.1.0
python-dotenv==0.19.0 